requirements.txt
asgiref==3.6.0
astroid==2.14.2
autopep8==2.0.1
backports.zoneinfo==0.2.1;python_version<"3.9"
dill==0.3.6
Django==4.1.6
django-filter==23.1
djangorestframework==3.14.0
dnspython==2.3.0
flake8==6.0.0
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
mypy==1.0.0
mypy-extensions==1.0.0
numpy==1.24.2
pandas==1.5.3
platformdirs==3.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
pylint==2.16.2
pylint-django==2.5.3
pylint-plugin-utils==0.7
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.2.4
tomli==2.0.1
tomlkit==0.11.6
typing-extensions==4.4.0
wrapt==1.14.1
